+++ doc/CHANGELOG.md
@@ -1,8 +1,12 @@
 
 ## Latest changes
 
+## Release 4.2.0
+
 #### :ghost: Maintenance
 * Update unstructured trajectory set calculation for pedestrians
+* Improved user interface of python binding
+* Suppressed python code generation build output
 
 ## Release 4.1.0